requirements.txt
sphinx
pymongo >= 3.0.2
python-magic
macholib
